Fix lift
* Changes:

Remodel the implementation of lift:

- With ManuallyDrop.
- With pointer reads/writes.

* Motivation:

In the URLO discussion, @steffanh found a counter-example which
triggered MIRI. Jasoncarr0 then iterated over the design until they
obtained an implementation which passed the Stacked Borrow model.

I took jasoncarr0's implementation and just applied a cosmetic change
-- wrapping the unsafe block into a function, for easy reuse -- as well
as added a number of comments to explain what was going on and justify
the safety of the scheme.

* Limitation:

The implementation is still considered experimental.